Fix type errors in wxString::Format()
wxWidgets-3.0.2 has strong type checking of arguments passed to wxString::Format.
If the passed type doesn't match the format string an exception is thrown.

+++ unittests/muleunit/testcase.cpp
@@ -60,7 +60,7 @@ const wxString& TestCase::getName() const
 
 bool TestCase::run()
 {
-	Print(wxT("\nRunning test-collection \"") + m_name + wxString::Format(wxT("\" with %u test-cases:"), m_tests.size()));
+	Print(wxT("\nRunning test-collection \"") + m_name + wxString::Format(wxT("\" with %u test-cases:"), static_cast<unsigned int>(m_tests.size())));
 
 	bool failures = false;
